+++ packages/ui/src/form/primitives/autocomplete/item.tsx
@@ -6,16 +6,12 @@ import resetStyles from '../../../utils/css/resets.module.css';
 import type { AutocompleteItemProps } from './types';
 
 export const Item = forwardRef< HTMLDivElement, AutocompleteItemProps >(
-	function Item(
-		{ className, children, size = 'default', ...restProps },
-		ref
-	) {
+	function Item( { className, children, ...restProps }, ref ) {
 		return (
 			<_Autocomplete.Item
 				className={ clsx(
 					resetStyles[ 'box-sizing' ],
 					itemPopupStyles.item,
-					size === 'small' && itemPopupStyles[ 'is-size-small' ],
 					className
 				) }
 				ref={ ref }