+++ packages/ui/src/form/primitives/combobox/item.tsx
@@ -9,13 +9,7 @@ import type { ComboboxItemProps } from './types';
 
 export const Item = forwardRef< HTMLDivElement, ComboboxItemProps >(
 	function Item(
-		{
-			className,
-			children,
-			size = 'default',
-			variant = 'default',
-			...restProps
-		},
+		{ className, children, variant = 'default', ...restProps },
 		ref
 	) {
 		return (
@@ -23,7 +17,6 @@ export const Item = forwardRef< HTMLDivElement, ComboboxItemProps >(
 				className={ clsx(
 					resetStyles[ 'box-sizing' ],
 					itemPopupStyles.item,
-					size === 'small' && itemPopupStyles[ 'is-size-small' ],
 					className
 				) }
 				ref={ ref }
@@ -36,7 +29,7 @@ export const Item = forwardRef< HTMLDivElement, ComboboxItemProps >(
 						variant !== 'creatable' &&
 							itemPopupStyles[ 'item-indicator-icon' ]
 					) }
-					size={ size === 'small' ? 20 : 24 }
+					size={ 24 }
 				/>
 				{ children }
 			</_Combobox.Item>