+++ packages/ui/src/form/primitives/combobox/types.ts
@@ -1,7 +1,6 @@
 import type { Combobox as _Combobox } from '@base-ui/react/combobox';
 import type { ReactElement } from 'react';
 import type { ComponentProps } from '../../../utils/types';
-import type { ItemPopupSize } from '../../../utils/item-popup-size';
 import type { InputLayoutProps } from '../input-layout/types';
 
 type ComboboxSize = Exclude< InputLayoutProps[ 'size' ], 'small' >;
@@ -61,12 +60,6 @@ export type ComboboxInputGroupProps = ComponentProps<
 
 export type ComboboxItemProps = ComponentProps< typeof _Combobox.Item > & {
 	children?: React.ReactNode;
-	/**
-	 * The size of the item.
-	 *
-	 * @default 'default'
-	 */
-	size?: ItemPopupSize;
 	/**
 	 * The variant of the item.
 	 *