+++ packages/ui/src/form/select-control/select-control.tsx
@@ -1,6 +1,5 @@
 import { forwardRef } from '@wordpress/element';
 import { Field, Select } from '../primitives';
-import { SelectControlSizeContext } from './context';
 import { Item } from './item';
 import type { SelectControlProps } from './types';
 
@@ -26,8 +25,6 @@ export const SelectControl = forwardRef<
 	},
 	ref
 ) {
-	const itemSize = size === 'small' ? 'small' : undefined;
-
 	return (
 		<Field.Root className={ className }>
 			<Field.Label hideFromVision={ hideLabelFromVision }>
@@ -42,20 +39,18 @@ export const SelectControl = forwardRef<
 					{ triggerContent }
 				</Select.Trigger>
 				<Select.Popup>
-					<SelectControlSizeContext.Provider value={ itemSize }>
-						{ children !== undefined
-							? children
-							: items?.map( ( item ) => (
-									<Item
-										key={ item.value ?? 'null' }
-										value={ item }
-										label={ item.label }
-										disabled={ item.disabled }
-									>
-										{ item.label }
-									</Item>
-							  ) ) }
-					</SelectControlSizeContext.Provider>
+					{ children !== undefined
+						? children
+						: items?.map( ( item ) => (
+								<Item
+									key={ item.value ?? 'null' }
+									value={ item }
+									label={ item.label }
+									disabled={ item.disabled }
+								>
+									{ item.label }
+								</Item>
+						  ) ) }
 				</Select.Popup>
 			</Select.Root>
 			{ description && (